+++ src/apps/Altinn.AccessManagement/test/AccessMgmt.Tests/AccessMgmt.Tests.csproj
@@ -15,6 +15,7 @@
     <ItemGroup>
         <ProjectReference Include="..\..\..\..\libs\Altinn.Authorization.Api.Contracts\src\Altinn.Authorization.Api.Contracts\Altinn.Authorization.Api.Contracts.csproj" />
         <ProjectReference Include="..\..\src\Altinn.AccessManagement\Altinn.AccessManagement.csproj" />
+        <ProjectReference Include="..\Altinn.AccessManagement.TestUtils\Altinn.AccessManagement.TestUtils.csproj" />
     </ItemGroup>
 
     <ItemGroup>
@@ -56,5 +57,9 @@
 
     <ItemGroup>
         <None Remove="Data\AuthorizedParties\TestDataAppsInstanceDelegation.cs" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Data\Xacml\3.0\ResourceRegistry\nav_sykepenger_dialog\" />
     </ItemGroup>
 </Project>
